change float -> double
* float is SINGLE precision, for float64 use DOUBLE!!!
* change default il size to 10
* delete commented code for halley and secant, move to secant_example.pyx
and halley_example,pyx
* make voltage, il, and diode parameters accessible as arguments to
bench_cython_newton
* in newton.pyx make TOL and MAXITER module level constants
* remove halley and secant code, and if-else condition blocks containing
their algorithms, so this is only for newton.
* remove tol and maxiter args, and remove validation checks for them
* remove myargs everywhere, can't use splat in cython
* remove trainling comments of old code
* remove comments about complex numbers and the 1.0 * x0 way to make it
float, b/c this statically typed, so it can only be double

+++ scipy/optimize/cython_optimize/zeros.pxd
@@ -1,4 +1,4 @@
 
-ctypedef float (*callback_type)(float, tuple);
+ctypedef double (*callback_type)(double, tuple);
 
-cdef float newton(callback_type func, float x0, callback_type fprime, tuple args)
+cdef double newton(callback_type func, double x0, callback_type fprime, tuple args)
